feat : SCRUM-21 Testcontainers + DynamicPropertySource 코드를 추상 클래스로 분리하고 상속 받아 사용하도록 수정
