+++ src/test/java/com/ixi_U/IxiUApplicationTests.java
@@ -1,32 +1,15 @@
 package com.ixi_U;
 
+import com.ixi_U.common.AbstractNeo4jContainer;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.Neo4jContainer;
-import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
 @SpringBootTest
 @ActiveProfiles("test")
 @Testcontainers
-class IxiUApplicationTests {
-
-    @Container
-    private static Neo4jContainer<?> neo4j = new Neo4jContainer<>(
-            DockerImageName.parse("neo4j:4.4"))
-            .withAdminPassword(System.getenv().getOrDefault("GRAPH_DB_PASSWORD", "testPassword"));
-
-    @DynamicPropertySource
-    static void overrideNeo4jProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.neo4j.uri", neo4j::getBoltUrl);
-        registry.add("spring.neo4j.authentication.username", () -> "neo4j");
-        registry.add("spring.neo4j.authentication.password",
-                () -> System.getenv().getOrDefault("GRAPH_DB_PASSWORD", "testPassword"));
-    }
+class IxiUApplicationTests extends AbstractNeo4jContainer {
 
     @Test
     void contextLoads() {
