+++ src/test/java/com/ixi_U/user/repository/UserAndPlanRepositoryTest.java
@@ -3,6 +3,7 @@ package com.ixi_U.user.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import com.ixi_U.common.AbstractNeo4jContainer;
 import com.ixi_U.common.exception.GeneralException;
 import com.ixi_U.common.exception.enums.PlanException;
 import com.ixi_U.common.exception.enums.UserException;
@@ -18,28 +19,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.neo4j.DataNeo4jTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.transaction.annotation.Transactional;
-import org.testcontainers.containers.Neo4jContainer;
-import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
 
 @Import(SubscribedService.class)
 @DataNeo4jTest
 @ActiveProfiles("test")
-@Transactional
 @Testcontainers
 @DisplayName("Testcontainers 기반 Neo4j DB와 실제 연동되는 구독관계 저장 테스트")
-class UserAndPlanRepositoryTest {
-
-    @Container
-    private static final Neo4jContainer<?> neo4j = new Neo4jContainer<>(
-            DockerImageName.parse("neo4j:5.15"))
-            .withAdminPassword(System.getenv().getOrDefault("GRAPH_DB_PASSWORD", "testPassword"))
-            .withReuse(true);
+class UserAndPlanRepositoryTest extends AbstractNeo4jContainer {
 
     @Autowired
     UserRepository userRepository;
@@ -47,14 +35,6 @@ class UserAndPlanRepositoryTest {
     PlanRepository planRepository;
     @Autowired
     SubscribedService subscribedService;
-
-    @DynamicPropertySource
-    static void overrideNeo4jProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.neo4j.uri", neo4j::getBoltUrl);
-        registry.add("spring.neo4j.authentication.username", () -> "neo4j");
-        registry.add("spring.neo4j.authentication.password",
-                () -> System.getenv().getOrDefault("GRAPH_DB_PASSWORD", "testPassword"));
-    }
 
     @Test
     @DisplayName("회원과 요금제 생성 후 실제 SUBSCRIBED 관계가 DB에 저장된다")
@@ -161,6 +141,5 @@ class UserAndPlanRepositoryTest {
                 .hasMessage(PlanException.ALREADY_SUBSCRIBED_PLAN.getMessage());
 
     }
-
 
 }
